Final Version
This is the Final version of my Medical Insurance Project that I have been actively working on. This includes my knowledge of all the data types in Python and so on. This takes data from a csv file, imports it,  then orgranizes, and analyzes the information in different ways. Please make sure that insurance.csv is included in the same folder that the project is opened in.